added imgui menu scope concept

+++ LigumX/src/ImguiHelpers.cpp
@@ -31,12 +31,12 @@ bool ImguiHelpers::ShowLong(LXString& name, long& value, long min, long max)
 	return ShowLong(name.c_str(), value, min, max);
 }
 
-bool ImguiHelpers::ShowIVec2(const char* name, glm::ivec2& value, float min, float max)
+bool ImguiHelpers::ShowIVec2(const char* name, glm::ivec2& value, int min, int max)
 {
 	return ImGui::DragInt2(name, (int*)&value, 1.f, min, max);
 }
 
-bool ImguiHelpers::ShowIVec2(LXString& name, glm::ivec2& value, float min, float max)
+bool ImguiHelpers::ShowIVec2(LXString& name, glm::ivec2& value, int min, int max)
 {
 	return ShowIVec2(name.c_str(), value, min, max);
 }
